requirements.txt
mysqlclient
django
djoser
djangorestframework
djangorestframework-simplejwt
python-dotenv
drf-nested-routers